Customized beam should be respected by nanoBragg.
Setting user_beam=True allows a dxtbx beam model or a generalized
user-defined beam direction to be locked in place without being reset
according to the conventions of a standard format.  Unless
user_beam is True, the pix0 (detector origin) and Fclose/Sclose (beam
center) vectors will be updated when init_beamcenter() calls
update_beamcenter().  This is an issue for tilted cameras.